requirements.txt
flask
Pillow
deep-translator
easyocr
torch
pyngrok
waitress
logging